server/python/requirements.txt
# OBS 方案依赖
obs-websocket-py
websocket-client
numpy
opencv-python
websockets
